+++ docker-compose.yml
@@ -1,51 +1,15 @@
 version: '3.1'
 services:
-  ewmstats:
-    build: ewmstats/server
-    container_name: ewmstats-container
+  stats-server:
     ports:
       - "9090:9090"
-    depends_on:
-      - stats-db
-    environment:
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://stats-db:5432/stats-db
-      - SPRING_DATASOURCE_USERNAME=stats
-      - SPRING_DATASOURCE_PASSWORD=stats
-
-  ewmservice:
-    build: ewmservice
-    container_name: ewmservice-container
-    ports:
-      - "8080:8080"
-    depends_on:
-      - ewm-db
-    environment:
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://ewm-db:5432/ewm-db
-      - SPRING_DATASOURCE_USERNAME=ewm
-      - SPRING_DATASOURCE_PASSWORD=ewm
 
   stats-db:
-    image: postgres:16.1
-    container_name: stats-db-container
-    volumes:
-      - /var/lib/postgresql/data/
+    image: postgres:14-alpine
+
+  ewm-service:
     ports:
-      - "6541:5432"
-    environment:
-      - POSTGRES_DB=stats-db
-      - POSTGRES_USER=stats
-      - POSTGRES_PASSWORD=stats
+      - "8080:8080"
 
   ewm-db:
-    image: postgres:16.1
-    container_name: ewn-db-container
-    volumes:
-      - /var/lib/postgresql/data/
-    ports:
-      - "6542:5432"
-    environment:
-      - POSTGRES_DB=ewm-db
-      - POSTGRES_USER=ewm
-      - POSTGRES_PASSWORD=ewm
-
-
+    image: postgres:14-alpine
