+++ pom.xml
@@ -12,7 +12,7 @@
 
 	<name>Explore With Me</name>
 	<modules>
-		<module>stats-server</module>
+		<module>stats-service</module>
         <module>ewm-service</module>
     </modules>
