+++ stats-service/client/src/main/java/ru/practicum/client/StatsClient.java
