+++ stats-service/dto/src/main/java/ru/practicum/dto/StatsDto.java
