+++ stats-service/dto/src/main/java/ru/practicum/utils/SimpleDateTimeFormatter.java
@@ -1,0 +1,20 @@
+package ru.practicum.utils;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class SimpleDateTimeFormatter {
+    public static final String PATTERN = "yyyy-MM-dd HH:mm:ss";
+
+    private SimpleDateTimeFormatter() {
+
+    }
+
+    public static String toString(LocalDateTime dateTime) {
+        return dateTime.format(DateTimeFormatter.ofPattern(PATTERN));
+    }
+
+    public static LocalDateTime parse(String dateTimeString) {
+        return LocalDateTime.parse(dateTimeString, DateTimeFormatter.ofPattern(PATTERN));
+    }
+}
