+++ stats-service/server/Dockerfile
