+++ stats-service/server/src/main/java/ru/practicum/EWMStatsApplication.java
