+++ stats-service/server/src/main/java/ru/practicum/server/controller/StatsController.java
