+++ stats-service/server/src/main/java/ru/practicum/server/mapper/StatsMapper.java
