+++ stats-service/server/src/main/java/ru/practicum/server/model/EndpointHit.java
