+++ stats-service/server/src/main/java/ru/practicum/server/service/StatsServiceImpl.java
