+++ stats-service/server/src/main/java/ru/practicum/server/utils/ErrorResponse.java
