+++ stats-service/server/src/main/resources/application-test.yaml
