+++ stats-service/server/src/test/resources/clean-up.sql
